+++ ints_test.go
@@ -62,95 +62,95 @@ func FuzzIntsFromBy(f *testing.F) {
 }
 
 func BenchmarkInts(b *testing.B) {
-	Ints[int]().Take(b.N).Consume()
+	Ints[int]().Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsFrom_negative(b *testing.B) {
-	IntsFrom(0).Take(b.N).Consume()
+	IntsFrom(0).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsFrom_zero(b *testing.B) {
-	IntsFrom(0).Take(b.N).Consume()
+	IntsFrom(0).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsFrom_positive(b *testing.B) {
-	IntsFrom(10000000).Take(b.N).Consume()
+	IntsFrom(10000000).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsBy_zero(b *testing.B) {
-	IntsBy(0).Take(b.N).Consume()
+	IntsBy(0).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsBy_smallIncreasing(b *testing.B) {
-	IntsBy(1).Take(b.N).Consume()
+	IntsBy(1).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsBy_bigIncreasing(b *testing.B) {
-	IntsBy(10000000).Take(b.N).Consume()
+	IntsBy(10000000).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsBy_smallDecreasing(b *testing.B) {
-	IntsBy(-1).Take(b.N).Consume()
+	IntsBy(-1).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsBy_bigDecreasing(b *testing.B) {
-	IntsBy(-10000000).Take(b.N).Consume()
+	IntsBy(-10000000).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsFromBy_negativeByZero(b *testing.B) {
-	IntsFromBy(-10000000, 0).Take(b.N).Consume()
+	IntsFromBy(-10000000, 0).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsFromBy_negativeBySmallIncreasing(b *testing.B) {
-	IntsFromBy(-10000000, 1).Take(b.N).Consume()
+	IntsFromBy(-10000000, 1).Take(uint(b.N)).Consume()
 }
 func BenchmarkIntsFromBy_negativeByLargeIncreasing(b *testing.B) {
-	IntsFromBy(-10000000, 10000000).Take(b.N).Consume()
+	IntsFromBy(-10000000, 10000000).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsFromBy_negativeBySmallDecreasing(b *testing.B) {
-	IntsFromBy(-10000000, -1).Take(b.N).Consume()
+	IntsFromBy(-10000000, -1).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsFromBy_negativeByLargeDecreasing(b *testing.B) {
-	IntsFromBy(-10000000, -10000000).Take(b.N).Consume()
+	IntsFromBy(-10000000, -10000000).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsFromBy_zeroByZero(b *testing.B) {
-	IntsFromBy(0, 0).Take(b.N).Consume()
+	IntsFromBy(0, 0).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsFromBy_zeroBySmallIncreasing(b *testing.B) {
-	IntsFromBy(0, 1).Take(b.N).Consume()
+	IntsFromBy(0, 1).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsFromBy_zeroByLargeIncreasing(b *testing.B) {
-	IntsFromBy(0, 10000000).Take(b.N).Consume()
+	IntsFromBy(0, 10000000).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsFromBy_zeroBySmallDecreasing(b *testing.B) {
-	IntsFromBy(0, -1).Take(b.N).Consume()
+	IntsFromBy(0, -1).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsFromBy_zeroByLargeDecreasing(b *testing.B) {
-	IntsFromBy(0, -10000000).Take(b.N).Consume()
+	IntsFromBy(0, -10000000).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsFromBy_positiveByZero(b *testing.B) {
-	IntsFromBy(10000000, 0).Take(b.N).Consume()
+	IntsFromBy(10000000, 0).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsFromBy_positiveBySmallIncreasing(b *testing.B) {
-	IntsFromBy(10000000, 1).Take(b.N).Consume()
+	IntsFromBy(10000000, 1).Take(uint(b.N)).Consume()
 }
 func BenchmarkIntsFromBy_positiveByLargeIncreasing(b *testing.B) {
-	IntsFromBy(10000000, 10000000).Take(b.N).Consume()
+	IntsFromBy(10000000, 10000000).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsFromBy_positiveBySmallDecreasing(b *testing.B) {
-	IntsFromBy(10000000, -1).Take(b.N).Consume()
+	IntsFromBy(10000000, -1).Take(uint(b.N)).Consume()
 }
 
 func BenchmarkIntsFromBy_positiveByLargeDecreasing(b *testing.B) {
-	IntsFromBy(10000000, -10000000).Take(b.N).Consume()
+	IntsFromBy(10000000, -10000000).Take(uint(b.N)).Consume()
 }
